go some commmon tools

+++ gintool/pager.go
@@ -4,14 +4,14 @@ import "math"
 
 //分页对象
 type Pager struct {
-	Page      int64   `form:"page"  json:"page"`           //当前页
-	PageSize  int64   `form:"pageSize"  json:"pageSize"`   //每页条数
-	Total     int64   `form:"total"  json:"total"`         //总条数
-	PageCount int64   `form:"pageCount"  json:"pageCount"` //总页数
-	NumStart  int64   `form:"numStart"  json:"numStart"`   //开始序数
+	Page      int   `form:"page"  json:"page"`           //当前页
+	PageSize  int   `form:"pageSize"  json:"pageSize"`   //每页条数
+	Total     int   `form:"total"  json:"total"`         //总条数
+	PageCount int   `form:"pageCount"  json:"pageCount"` //总页数
+	NumStart  int   `form:"numStart"  json:"numStart"`   //开始序数
 }
 
-func CreatePager(page, pagesize int64) *Pager {
+func CreatePager(page, pagesize int) *Pager {
 	if page < 1 {
 		page = 1
 	}
@@ -29,7 +29,7 @@ func (p *Pager) setNumStart() {
 	p.NumStart = (p.Page - 1) * p.PageSize
 }
 
-func (p *Pager) SetTotal(total int64) {
+func (p *Pager) SetTotal(total int) {
 	p.Total = total
-	p.PageCount =int64(math.Ceil(float64(total) / float64(p.PageSize)))
+	p.PageCount =int(math.Ceil(float64(total) / float64(p.PageSize)))
 }
